server/requirements.txt
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
eventlet==0.25.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.1.1
Flask-SocketIO==4.3.0
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
paho-mqtt==1.5.0
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
